database sync functionality moved to main
